Weather app finished
Changes:
- New feature: The information button now opens up an information box
- New feature: Change between Farenhegiht and Celcius with the button on the top right
- New styles for both buttons above
- Improved styling for short devices

The app is complete. I could have improved on many things such as making the f/c button to change the city card elements already present in the UI or adding a delete button to every card element. However, I am moving on to a different project. All works correctly and I am happy with the results! Thank you for checking this project out. It was made to sharpen my skills in Vanilla JavaScript and with XHR objects, API calls, DOM manipulation, and JSON manipulation.